Migrate existing files to new structure
