+++ process_mining/src/conformance/object_centric/mod.rs
@@ -1,0 +1,1 @@
+pub mod object_centric_language_abstraction;
