+++ process_mining/src/core/event_data/case_centric/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod activity_projection;
 #[cfg(feature = "log-splitting")]
 pub mod event_log_splitter;
+pub mod partial_orders;
