+++ process_mining/src/core/process_models/case_centric/dfg/mod.rs
@@ -1,0 +1,2 @@
+pub mod dfg_struct;
+pub mod image_export;
