+++ process_mining/src/core/process_models/case_centric/mod.rs
@@ -1,1 +1,3 @@
+pub mod dfg;
 pub mod petri_net;
+pub mod process_tree;
