+++ process_mining/src/core/process_models/case_centric/process_tree/mod.rs
@@ -1,0 +1,1 @@
+pub mod process_tree_struct;
