+++ process_mining/src/core/process_models/object_centric/mod.rs
@@ -1,2 +1,3 @@
 pub mod oc_declare;
+pub mod ocdfg;
 pub mod ocpt;
