+++ process_mining/src/core/process_models/object_centric/ocdfg/mod.rs
@@ -1,0 +1,1 @@
+pub mod object_centric_dfg_struct;
