--- process_mining/src/object_centric/conformance/mod.rs
@@ -1,2 +1,0 @@
-/// Object-centric OCEL and OCPT abstraction and conformance checking
-pub mod object_centric_language_abstraction_struct;
