--- process_mining/src/object_centric/mod.rs
@@ -1,6 +1,0 @@
-/// Object-centric conformance checking
-pub mod conformance;
-/// Object-centric directly-follows graphs
-pub mod object_centric_dfg_struct;
-/// Object-centric process trees
-pub mod ocpt;
